% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohby.r
\name{expand}
\alias{expand}
\title{Expand an 'OhBy' URL}
\usage{
expand(ohby_url)
}
\arguments{
\item{ohby_url}{the URL to expand}
}
\description{
Expand an 'OhBy' URL
}
\examples{
expand("https://0x.co/3BE87E")
expand("https://0x.co/6VZ69Z")
}

