% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohby-package.R
\docType{package}
\name{ohby}
\alias{ohby}
\alias{ohby-package}
\title{Tools to Work with the 'OhBy' Shortening Service}
\description{
Tools to Work with the 'OhBy' Shortening Service
}
\author{
Bob Rudis (@hrbrmstr)
}

