% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohby.r
\name{shorten}
\alias{shorten}
\title{Create an "OhBy" code for some content}
\usage{
shorten(content, expiration = "never", visibility = c("public", "private"),
  redirect = FALSE)
}
\arguments{
\item{content}{up to 4096 bytes of content, optionally a URL}

\item{expiration}{either "\code{never}" or a number followed by one of \code{m h D W M Y} for
\emph{minute}, \emph{hour}, \emph{Day}, \emph{Week}, \emph{Month}, \emph{Year}.}

\item{visibility}{either "\code{public}" or "\code{private}". The \emph{public} codes are
indexable by search engines}

\item{redirect}{if \code{content} is a URL, then set this to \code{TRUE} to have it behave like a URL
shortenter.}
}
\description{
What is "oh by"? "oh by" is a simple primitive that you can use to condense information.
It is, sort of, a combination of a pastebin-style site and a URL shortener.
The real utility are the easily recognizable codes, prefixed with "0x".
}
\examples{
shorten("http://rud.is/b")
shorten("The quick, brown fox jumps over the lazy dog")
}

